requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
beautifulsoup4==4.9.1
bottlenose==1.1.8
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.1.0
certifi==2020.4.5.2
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
contextlib2==0.6.0
crayons==0.3.1
cssselect==1.1.0
dateparser==0.7.6
distlib==0.3.0
distro==1.4.0
dnspython==2.0.0
et-xmlfile==1.0.1
fake-useragent==0.1.11
firebase-admin==4.3.0
Flask==1.1.2
flask-paginate==0.7.0
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-firestore==1.7.0
google-cloud-storage==1.29.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
GoogleNews==1.3.8
GoogleScraper==0.2.4
grpcio==1.29.0
html5lib==1.0.1
httplib2==0.18.1
idna==2.9
ipaddr==2.2.0
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
lockfile==0.12.2
lxml==4.5.1
MarkupSafe==1.1.1
msgpack==1.0.0
multidict==4.7.6
newsapi-python==0.2.6
numpy==1.18.5
oauthlib==3.1.0
openpyxl==3.0.5
packaging==20.3
pandas==1.0.4
pep517==0.8.2
Pillow==7.2.0
progress==1.5
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pygeohash==1.2.0
pymongo==3.11.0
PyMySQL==0.9.3
pyparsing==2.4.6
python-amazon-simple-product-api==2.2.11
python-dateutil==2.8.1
python-http-client==3.3.1
pytoml==0.1.21
pytz==2020.1
regex==2020.9.27
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.6
selenium==3.141.0
sendgrid==6.4.7
six==1.15.0
slack-sdk==3.0.0
soupsieve==2.0.1
SQLAlchemy==1.3.18
starkbank-ecdsa==1.1.0
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.25.9
webdriver-manager==3.2.1
webencodings==0.5.1
Werkzeug==1.0.1
yarl==1.4.2
youtube-data-api==0.0.20
youtube-dl==2020.6.16.1
